refactor: split session.rs into session/ module with types and test separation
- Convert monolithic session.rs (2,493 lines) into session/ directory
- Extract types.rs (~300 lines): ConversionMode, SessionState, Composition,
  CandidateState, StabilityTracker, FrozenPrefix, MarkedText, CandidateAction,
  KeyResponse, AsyncCandidateRequest, SideEffects, AsyncGhostRequest,
  key codes, constants, and utility functions
- Split tests (~1,090 lines) into tests/ subdirectory:
  basic.rs, candidates.rs, submode.rs, auto_commit.rs, ghost.rs
- Logic remains in mod.rs (to be split in next PR)
- No FFI changes required; public API re-exported from mod.rs
